pacman-hook-kernel-install: Skip when vmlinuz doesn't exist.
When updating the kernel, the previous version may get added to
version=() because extramodules also triggers when removed. To avoid
errors such as:

    Failed to open kernel image file '/usr/lib/modules/7.0.0-1-cachyos/vmlinuz': No such file or directory

Check for existence of vmlinuz, and skip if it is not found.

+++ pacman-hook-kernel-install/kernel-install.sh
@@ -32,6 +32,7 @@ if ((all_kernels)); then
 else
     for kver in "${!versions[@]}"; do
 	kimage="/usr/lib/modules/$kver/vmlinuz"
+	[[ ! -e "$kimage" ]] && continue
 	kernel-install "$@" "$kver" "$kimage" || true
     done
 fi
